Update Notebook
- Improved backup timer on Trainer. Now it count from start Colab runtime.

- Added "Manage workspace" block.
You can export to and from the drive, both the entire workspace and its individual parts. For example, if you need to quickly load aligned images, sort them and load them back. Now you do not need to export all the workspace to the drive.

Also, you can delete and re-create, like all workspace or parts of it, if something went wrong.